Conflicts:
	.gitignore
	api/wrapper/xwrap.c
	api/xsocket/Xaccept.c
	api/xsocket/XgetDAGbyName.c
	api/xsocket/Xgetaddrinfo.c
	api/xsocket/Xutil.h
	api/xsocket/state.c
	applications/tinyhttpd/httpd.c
	applications/tinyhttpd/simpleclient.c
	bin/make_sockconf
	bin/xconfig.py
	bin/xianet
	click-2.0.1/conf/xia_router_lib.click
	click-2.0.1/elements/xia/xiaxidtypecounter.cc
	click-2.0.1/elements/xia/xroute.cc
	click-2.0.1/elements/xia/xtransport.cc
	daemons/Makefile
	daemons/xhcp/xhcp_client.cc
	daemons/xrouted/xrouted.cc
	daemons/xrouted/xrouted.hh
	etc/click/xia_address.click
	etc/click/xia_local_topology.click
	etc/hosts.xia.example
	tools/visualizations/distributed_click/Visualizer.swf
	tools/visualizations/distributed_click/source/edu/cmu/topology/Topology.as

+++ api/xssl/XSSL_accept.c
@@ -45,7 +45,7 @@ int XSSL_accept(XSSL *xssl) {
 
 	/* Send SERVER HELLO */
 	sprintf(buf, "SERVER HELLO");
-	if ((n = Xsend(xssl->sockfd, buf, strlen(buf), 0)) != strlen(buf)) {
+	if ((n = Xsend(xssl->sockfd, buf, strlen(buf), 0)) != (int)strlen(buf)) {
 		ERROR("ERROR sending SERVER HELLO");
 		return 0;
 	}
@@ -76,17 +76,17 @@ int XSSL_accept(XSSL *xssl) {
 	offset += siglen;
 
 	// Send it!
-	if ((n = Xsend(xssl->sockfd, buf, offset, 0)) != offset) {
+	if ((n = Xsend(xssl->sockfd, buf, offset, 0)) != (int)offset) {
 		ERROR("ERROR sending public keys");
 		return 0;
 	}
 
-	DBGF("Sent keys (%d bytes)\n\tkeylen: %d\n\ttempkeylen: %d\n\tsiglen:%d", offset, keybufsize, tempkeybufsize, siglen);
+	DBGF("Sent keys (%u bytes)\n\tkeylen: %u\n\ttempkeylen: %u\n\tsiglen:%u", (unsigned)offset, (unsigned)keybufsize, (unsigned)tempkeybufsize, (unsigned)siglen);
 
 
 	/* Send SERVER DONE */
 	sprintf(buf, "SERVER DONE");
-	if ((n = Xsend(xssl->sockfd, buf, strlen(buf), 0)) != strlen(buf)) {
+	if ((n = Xsend(xssl->sockfd, buf, strlen(buf), 0)) != (int)strlen(buf)) {
 		ERROR("ERROR sending SERVER DONE");
 		return 0;
 	}
